New architecture release v.0
