+++ layouter.py
@@ -1,7 +1,7 @@
 import copy
 import numpy as np
 from numpy import arctan2, pi
-from mol2_chain_q import atoms_and_bonds,  bonds_of_paired, Notation
+from mol2_chain_q import atoms_and_bonds,  bonds_of_paired
 from mol2_worker import xyz_names_bonds
 
 
@@ -120,31 +120,34 @@ if __name__ == '__main__':
         n_param = 6
 
         # bs, ass = xyz_names_bonds(name + '.mol2')
-        atoms_info = atoms_and_bonds(file_name + '.mol2')
-        ln = Notation(n=n_param, info_from_file=xyz_names_bonds(file_name + '.mol2'))
-        paired = bonds_of_paired(ln.bonds)
-        dim_structure = dimensional_structure(ln, relax=True)
-        write_mol2_file('My_' + name + '_' + 'q0.mol2', atoms_info, dim_structure, bonds=paired)
 
-        file_name_new = 'My_' + name + '_' + 'q0'
-        atoms_info2 = atoms_and_bonds(file_name_new + '.mol2')
-        ln2 = Notation(n=n_param, info_from_file=xyz_names_bonds(file_name_new + '.mol2'))
 
-        flags = []
-        for i in ln.notation.keys():
-            flags.append(ln.notation[i].sort()==ln2.notation[i].sort())
-            if ln.notation[i].sort() != ln2.notation[i].sort():
-                print(ln.notation[i], ln2.notation[i])
-        print(name_sh, len(flags) - sum(flags), 'errors')
+        # atoms_info = atoms_and_bonds(file_name + '.mol2')
+        # ln = Notation(n=n_param, info_from_file=xyz_names_bonds(file_name + '.mol2'))
+        # paired = bonds_of_paired(ln.bonds)
+        # dim_structure = dimensional_structure(ln, relax=True)
+        # write_mol2_file('My_' + name + '_' + 'q0.mol2', atoms_info, dim_structure, bonds=paired)
+        #
+        # file_name_new = 'My_' + name + '_' + 'q0'
+        # atoms_info2 = atoms_and_bonds(file_name_new + '.mol2')
+        # ln2 = Notation(n=n_param, info_from_file=xyz_names_bonds(file_name_new + '.mol2'))
+        #
+        # flags = []
+        # for i in ln.notation.keys():
+        #     flags.append(ln.notation[i].sort()==ln2.notation[i].sort())
+        #     if ln.notation[i].sort() != ln2.notation[i].sort():
+        #         print(ln.notation[i], ln2.notation[i])
+        # print(name_sh, len(flags) - sum(flags), 'errors')
+        #
+        # flags = []
+        # for i in ln.bonds.keys():
+        #     b1, b2 = sorted(ln.bonds[i]), sorted(ln2.bonds[i])
+        #     for j in range(len(b1)):
+        #         flags.append(b1[j][1] == b2[j][1])
+        #         if b1[j][1] != b2[j][1]:
+        #             print(b1[j][1], b2[j][1])
+        # print(name_sh, len(flags) - sum(flags), 'length errors')
 
-        flags = []
-        for i in ln.bonds.keys():
-            b1, b2 = sorted(ln.bonds[i]), sorted(ln2.bonds[i])
-            for j in range(len(b1)):
-                flags.append(b1[j][1] == b2[j][1])
-                if b1[j][1] != b2[j][1]:
-                    print(b1[j][1], b2[j][1])
-        print(name_sh, len(flags) - sum(flags), 'length errors')
         # print(ln.notation == ln2.notation)
 
         # dim_structure = dimensional_structure(ln2, relax=True)
