+++ scr_old/colors_from_plots.py
