+++ scr_old/gradient_descent_rotation.py
