+++ scr_old/gradient_plotter.py
