+++ scr_old/mol2_zero_bonds.py
