+++ scr_old/mopac_bonder.py
