+++ scr_old/separeted_opt.py
