+++ scr_old/two_states_to_reaction.py
